+++ cobaya/likelihoods/gaussian_mixture/__init__.py
@@ -1,2 +1,2 @@
-from .gaussian_mixture import gaussian_mixture, random_mean, random_cov, \
+from .gaussian_mixture import GaussianMixture, random_mean, random_cov, \
     info_random_gaussian_mixture