+++ cobaya/typing.py
@@ -5,7 +5,7 @@ import sys
 
 # an immutable empty dict (e.g. for argument defaults)
 empty_dict = MappingProxyType({})
-# empty list-type object that can be used like None to test if assigned
+# empty iterator-compatible object that can be used like None to test if assigned
 unset_params: Sequence[str] = ()
 
 InfoDict = Dict[str, Any]
@@ -19,6 +19,7 @@ TheoriesDict = Dict[str, Union[None, TheoryDict, Type]]
 LikesDict = Dict[str, Union[None, LikeDict, Type]]
 SamplersDict = Dict[str, Optional[SamplerDict]]
 PriorsDict = Dict[str, Union[str, Callable]]
+
 # parameters in a params list can be specified on input by
 # 1. a ParamDict dictionary
 # 2. constant value
@@ -27,8 +28,8 @@ PriorsDict = Dict[str, Union[str, Callable]]
 # 5. Sequence specifying uniform prior range [min, max] and optionally
 #    'ref' mean and standard deviation for starting positions, and optionally
 #    proposal width. Allowed lengths, 2, 4, 5
-ParamSpec = Union['ParamDict', None, str, float, Sequence[float]]
-ParamsDict = Dict[str, ParamSpec]
+ParamInput = Union['ParamDict', None, str, float, Sequence[float]]
+ParamsDict = Dict[str, ParamInput]
 ExpandedParamsDict = Dict[str, 'ParamDict']
 
 partags = {"prior", "ref", "proposal", "value", "drop",
@@ -101,8 +102,8 @@ if sys.version_info >= (3, 8):
 
 else:
     # avoid PyCharm parsing these too...
-    for _x in ('InputDict', 'ParamDict', 'ModelDict', 'PostDict'):
-        globals()[_x] = InfoDict
+    globals().update((k, InfoDict) for k in
+                     ('InputDict', 'ParamDict', 'ModelDict', 'PostDict'))
     globals()['LiteralFalse'] = bool
 
 try: